# backend/src/main/java/com/assessment/securedhelloworld/admin/AdminUserController.java
package com.assessment.securedhelloworld.admin;

import com.assessment.securedhelloworld.auth.AppUserDetails;
import com.assessment.securedhelloworld.user.Role;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

/**
 * REST surface for {@link AdminUserService} — gated by
 * {@code hasRole("ADMIN")} in {@code SecurityConfig}, not by any
 * annotation on this class.
 */
@RestController
@RequestMapping("/api/admin/users")
@org.springframework.security.access.prepost.PreAuthorize("hasRole('ADMIN')")
public class AdminUserController {

    private final AdminUserService adminUserService;

    public AdminUserController(AdminUserService adminUserService) {
        this.adminUserService = adminUserService;
    }

    @GetMapping
    public List<AdminUserView> listUsers() {
        return adminUserService.listUsers();
    }

    @PatchMapping("/{id}/status")
    public ResponseEntity<Void> updateStatus(
            @PathVariable Long id,
            @Valid @RequestBody UpdateStatusRequest request,
            @AuthenticationPrincipal AppUserDetails actingAdmin) {
        adminUserService.updateEnabled(id, request.getEnabled(), actingAdmin);
        return ResponseEntity.ok().build();
    }

    @PatchMapping("/{id}/role")
    public ResponseEntity<Void> updateRole(
            @PathVariable Long id,
            @Valid @RequestBody UpdateRoleRequest request,
            @AuthenticationPrincipal AppUserDetails actingAdmin) {
        adminUserService.updateRole(id, Role.valueOf(request.getRole()), actingAdmin);
        return ResponseEntity.ok().build();
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> deleteUser(
            @PathVariable Long id,
            @AuthenticationPrincipal AppUserDetails actingAdmin) {
        adminUserService.deleteUser(id, actingAdmin);
        return ResponseEntity.ok().build();
    }
}