# backend/src/main/java/com/assessment/securedhelloworld/auth/LoginRequest.java
package com.assessment.securedhelloworld.auth;

import jakarta.validation.constraints.NotBlank;

/**
 * Request body for {@code POST /api/login}.
 */
public class LoginRequest {

    @NotBlank
    private String username;

    @NotBlank
    private String password;

    public LoginRequest() {
        // Jackson
    }

    public LoginRequest(String username, String password) {
        this.username = username;
        this.password = password;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
}